+++ src/main/java/com/ejoongseok/wmslive/location/domain/LocationRepository.java
@@ -1,0 +1,20 @@
+package com.ejoongseok.wmslive.location.domain;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+public class LocationRepository {
+    private final Map<Long, Location> locations = new HashMap<>();
+    private Long sequence = 1L;
+
+    public void save(final Location location) {
+        location.assignId(sequence++);
+        locations.put(location.getLocationNo(), location);
+    }
+
+    public List<Location> findAll() {
+        return new ArrayList<>(locations.values());
+    }
+}
