+++ src/main/java/com/ejoongseok/wmslive/location/domain/StorageType.java
@@ -1,0 +1,11 @@
+package com.ejoongseok.wmslive.location.domain;
+
+public enum StorageType {
+    TOTE("토트 바구니");
+
+    private final String description;
+
+    StorageType(final String description) {
+        this.description = description;
+    }
+}
