+++ src/main/java/com/ejoongseok/wmslive/location/domain/UsagePurpose.java
@@ -1,0 +1,11 @@
+package com.ejoongseok.wmslive.location.domain;
+
+public enum UsagePurpose {
+    MOVE("이동");
+
+    private final String description;
+
+    UsagePurpose(final String description) {
+        this.description = description;
+    }
+}
